+++ docs/release-notes.md
@@ -14,6 +14,7 @@ hide:
 - Revisioning with `auto_now` as primary key.
 - Fix crash with `auto_now_add` as primary key.
 - Fix "$defs" not matching def names. This fixes problems with enums which appear multiple times and reduces requests.
+- Fix using ForeignKeys in embedded models/CompositeFields when the embedded model has no registry set (e.g. abstract).
 
 ## 0.32.4